refactor: #49 구인 글 카드 컴포넌트화
