+++ src/components/index.ts
@@ -85,6 +85,7 @@ import ApplyInfomation from '../pages/recruit/RecruitDetailPage/steps/ApplyInfom
 import ApplyInput from '../pages/recruit/RecruitDetailPage/steps/ApplyInput';
 import ApplySubmit from '../pages/recruit/RecruitDetailPage/steps/ApplySubmit';
 import Content from './meeteam/Content';
+import RecruitCardSmall from './meeteam/card/RecruitCardSmall';
 
 export {
 	Header,
@@ -165,4 +166,5 @@ export {
 	informationList,
 	role,
 	CONTENT,
+	RecruitCardSmall,
 };
