+++ src/pages/recruit/RecruitDetailPage/RecruitDetailPage.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useState } from 'react';
 import S from './RecruitDetailPage.styled';
 import {
 	Tag,
